Additional support for event bus reply failures and timeouts. Fixes #13.
There are now 3 parameters for an event bus handler. 1) The incoming
message listen for messages on; 2) The replier object; and 3) a 'meta'
object of sorts, that for now houses a function `fail()` that can be
used to signal a failure on the part of the recipient, and an `address`
property that contains the sender's address. An example:

    // A handler that always fails
    bus.registerHandler('some-address', function(msg, replier, meta) {
      meta.fail(23, 'Cannot calculate any more on this address: ' +
      meta.address);
    });