Use `handler.call()` instead of `handler()` on read streams.
This is primarily for compatibility with vertx/parse_tools.Parser
which needs to act like both an object and a function, and to do
so, overrides `Object.prototype.call`. None of this is really ideal
and the API should probably be re-evaluted for the next major release.

This change should have been included in 235bd16 but was overlooked.